Add query_timeout
For a given connection, requests are sent to the server in the order they are
seen in. But only when a request with a callback is sent, any subsequent request
will be queued in the client until the request completes. This is so that the
query_timeout is applied to only the query is was meant for.

Otherwise is a query is sent after a previous long-running query it could get
a timeout before the server has even started processing it

a query operation will return a guard object when not in a void context. If
this guard object goes out of scope, the query will be cancelled. This can
be used with a timer if an absolute timeout is needed